models.py
import torch
import torch.nn as nn
import numpy as np
from sklearn.preprocessing import MinMaxScaler
from openai import OpenAI
import os
from dotenv import load_dotenv
import apis

load_dotenv()
client = OpenAI(base_url="https://api.x.ai/v1", api_key=os.getenv('XAI_API_KEY'))

class LSTMModel(nn.Module):
    # (same as before)

def predict_price(data_df, column='price'):
    # (improved 50-epoch training as before)

def simulate_profitable_trade(query, pred_price, current_price, is_premium):
    if not is_premium:
        return "Premium unlocks real simulated profits!"
    profit_pct = ((pred_price - current_price) / current_price) * 100
    if 'arbitrage' in query.lower() or profit_pct > 5:
        return f"✅ SIMULATED TRADE EXECUTED!\nBought 1 ETH @ ${current_price:.2f}\nSold at prediction @ ${pred_price:.2f}\n**+{profit_pct:.1f}% PROFIT** (${(pred_price-current_price):.2f} gain)"
    return f"Simulated trade ready — projected +{profit_pct:.1f}%"

def grok_predict(...):  # unchanged