requirements.txt
flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
torch==2.3.1
pykalshi==1.0.0
stripe==8.7.0
flask-login==0.6.3
cryptography==43.0.0